requirements.txt
# Jarvis Voice Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Core audio processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0

# Speech-to-Text
openai-whisper>=20231117
# Alternative STT options:
# vosk>=0.3.45
# deepgram-sdk>=3.0.0

# Text-to-Speech
requests>=2.28.0
# For ElevenLabs (default)
# elevenlabs>=0.2.0  # Optional, we use direct API calls

# For local TTS alternatives:
# piper-tts>=1.0.0
# TTS>=0.17.0  # Coqui TTS

# For MP3 playback
pydub>=0.25.1

# LLM Providers
anthropic>=0.18.0
# openai>=1.0.0  # If using OpenAI LLM/TTS

# Wake Word Detection
pvporcupine>=3.0.0
# openwakeword>=0.5.0  # Alternative, no API key needed

# Home Assistant Integration (optional)
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
